Add license files
Licensing is a mess. The wrapper gem is MIT-licensed,
and Highcharts itself is either the Highsoft license
or it's available under Attribution-NonCommercial (CC BY-NC).

Additionally, the Highsoft website refers to
"under a Creative Commons (CC) Attribution-NonCommercial licence",
but it links to version 3.0 of the license. Unsure if it means
3.0 and later, or just 3.0.

This fixes #14